move folder queueOrderJobs to main tables
